Touch Typing Application v.05
NOTICE: This was build with the latest version of unity at the time
(5.1.1).

This is the base application which should give you enough functionality
to let you start working on getting the leap gestures working.
It spawns characters on either side of the players screen and they move
left or right.
The player can press the key and it will destroy the object. The
characters can also be destroyed if they hit a trigger around the
camera.
Characters can not be duplicated and will not reappear until the
original has been destroyed.
The code should have enough comments to let you figure out how to tweak
it.