Kody-Rift Edition v1
The changes begin preparing the game for rift integration.
These include:
-Changed asteroids spawning so that they will be un-parented from the
camera after setting their initial position. This has the effect that
they will always spawn just outside the cameras FoV but wont follow the
camera around.
-Changed asteroid de-spawning. Now asteroids will only de-spawn if hit
by a projectile OR if they have been alive for 12 seconds and they are
no longer visible to the renderer (ie outside camera view).
-Not a change but to describe how the player interaction works the
camera tracks the players head and they can rotate freely on the spot.
As the spaceship is the child of the camera the spaceship follows this
movement.
Rift Mode Requirements:
-Oculus Runtime v0.6.0.0. Set it up and make sure you can run the desk
demo (not crucial but a good way to check the rift works).
-In the unity editor go to the top bar and Edit->Project
Settings->Player and tick Virtual Reality Supported.
Other Changes:
-Fixed an issue with / not working when it was the only char in the
string. (Unity could not find an object named / and the projectile would
not move)
-Right shift can now be used for uppercase characters
-Reset spaceship rotation to 0,0,0 from 0,45,0. Works better with new
asteroid spawning.
Known Issues:
-If a player presses down both shift buttons and presses / then the
asteroid wont fire a projectile for any combination of those chars. The
char can only be destroyed by the default method.